+++ types/orchestrationTypes.d.ts
@@ -1,6 +1,6 @@
 import { FunctionOptions, FunctionTrigger, InvocationContext, LogHandler } from "@azure/functions";
-import { Task, TimerTask, TokenSource } from ".";
-import { EntityId, RetryOptions } from "../classes";
+import { RetryOptions, Task, TimerTask, TokenSource } from ".";
+import { EntityId } from "./entityTypes";
 
 /**
  * Type of a Generator that can be registered as an orchestration
